+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/configurations/ApplicationAsyncIncidents.java
@@ -17,6 +17,7 @@ import java.util.concurrent.Executor;
 
 import static io.tech1.framework.domain.properties.utilities.PropertiesAsserter.assertProperties;
 import static io.tech1.framework.domain.utilities.processors.ProcessorsUtility.getHalfOfCores;
+import static io.tech1.framework.domain.utilities.processors.ProcessorsUtility.getNumOfCores;
 
 @Configuration
 @EnableAsync
@@ -40,6 +41,7 @@ public class ApplicationAsyncIncidents implements AsyncConfigurer {
         var threadNamePrefix = this.applicationFrameworkProperties.getAsyncConfigs().getThreadNamePrefix();
         var taskExecutor = new ThreadPoolTaskExecutor();
         taskExecutor.setCorePoolSize(getHalfOfCores());
+        taskExecutor.setMaxPoolSize(getNumOfCores());
         taskExecutor.setThreadNamePrefix(threadNamePrefix);
         taskExecutor.setRejectedExecutionHandler(this.rejectedExecutionHandlerPublisher);
         taskExecutor.initialize();